backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.10.1
python-jose[cryptography]==3.4.0
passlib==1.7.4
pydantic[email]==2.11.7
python-dotenv==1.1.1
joblib==1.5.1
numpy==2.3.1
scikit-learn==1.8.0
langchain==0.3.27
langchain-mistralai==0.2.10
langdetect==1.0.9
deep-translator==1.11.4
